+++ R/distribution.R
@@ -14,8 +14,12 @@ NULL
 #'
 #' Additionally, it is possible to specify a \code{"limit"} to any predictions conducted on
 #' the background. This can be for instance a buffered layer by a certain dispersal distance (Cooper and Soberon, 2018)
-#' or a categorical layer representing biomes or soil conditions. See also the
-#' frequently asked question (FAQ) section on the homepage for more information.
+#' or a categorical layer representing biomes or soil conditions.
+#' Another option is to create a constraint by constructing a minimum convex polygon (MCP) using
+#' the supplied biodiversity data. This option can be enabled by setting
+#' \code{"limits_method"} to \code{"mcp"}. It is also possible to provide a small buffer
+#' to constructed MCP that way.
+#' See the frequently asked question (FAQ) section on the homepage for more information.
 #'
 #' See **Details** for a description of the internal functions available
 #' to modify or summarize data within the created object.
@@ -27,6 +31,12 @@ NULL
 #' [`SpatRaster`], [`sf`] or [`extent`] object.
 #' @param limits A [`SpatRaster`] or [`sf`] object that limits the prediction surface when
 #' intersected with input data (Default: \code{NULL}).
+#' @param limits_method A [`character`] of the method used for hard limiting a projection.
+#' Available options are \code{"none"} (Default), \code{"zones"} or \code{"mcp"}.
+#' @param mcp_buffer A [`numeric`] distance to buffer the mcp (Default \code{0}). Only used if
+#'  \code{"mcp"} is used.
+#' @param limits_clip [`logical`] Should the limits clip all predictors before fitting
+#' a model (\code{TRUE}) or just the prediction (\code{FALSE}, default).
 #'
 #' @details
 #' This function creates a [`BiodiversityDistribution-class`] object that in itself contains
@@ -83,18 +93,21 @@ NULL
 #' @export
 methods::setGeneric("distribution",
                     signature = methods::signature("background"),
-                    function(background, limits = NULL) standardGeneric("distribution"))
+                    function(background, limits = NULL, limits_method = "none", mcp_buffer = 0,limits_clip = FALSE) standardGeneric("distribution"))
 
 #' @name distribution
-#' @usage \S4method{distribution}{SpatRaster, ANY}(background, limits)
+#' @usage \S4method{distribution}{SpatRaster, ANY, character, numeric, logical}(background, limits, limits_method, mcp_buffer,limits_clip)
 #' @rdname distribution
 methods::setMethod(
   "distribution",
   methods::signature(background = "SpatRaster"),
-  function(background, limits = NULL) {
+  function(background, limits = NULL, limits_method = "none", mcp_buffer = 0,limits_clip = FALSE) {
     assertthat::assert_that(!missing(background) || !exists('background'),
                             inherits(limits,'SpatRaster') || inherits(limits, 'sf') || inherits(limits, 'Spatial') || is.null(limits),
-                            msg = 'No background file supplied!')
+                            is.character(limits_method),
+                            is.numeric(mcp_buffer),
+                            is.logical(limits_clip),
+                            msg = 'No background file supplied or limits misspecified!')
     # Check that arguments are valid
     assertthat::assert_that( inherits(background,'SpatRaster')  )
 
@@ -104,30 +117,39 @@ methods::setMethod(
     )
 
     # Rerun the distribution call with the object
-    distribution(newbg, limits)
+    distribution(newbg, limits, limits_method, mcp_buffer, limits_clip)
   })
 
 #' @name distribution
-#' @usage \S4method{distribution}{sf, ANY}(background, limits)
+#' @usage \S4method{distribution}{sf, ANY, character, numeric, logical}(background, limits, limits_method, mcp_buffer, limits_clip)
 #' @rdname distribution
 methods::setMethod(
   "distribution",
   methods::signature(background = "sf"),
-  function(background, limits = NULL) {
+  function(background, limits = NULL, limits_method = "none", mcp_buffer = 0, limits_clip = FALSE) {
     # Check that arguments are valid
     assertthat::assert_that(!missing(background) || !exists('background'),
                             inherits(limits,'SpatRaster') || inherits(limits, 'sf') || inherits(limits, 'Spatial') || is.null(limits),
+                            is.character(limits_method),
+                            is.numeric(mcp_buffer),
+                            is.logical(limits_clip),
                             msg = 'No background file supplied!')
     assertthat::assert_that(
       inherits(background,'sf'),
       unique(st_geometry_type(background)) %in% c('MULTIPOLYGON','POLYGON')
     )
 
+    # Small checks on alternative limit functionalities
+    limits_method <- match.arg(limits_method, c("none","zones", "mcp"), several.ok = FALSE)
+    assertthat::assert_that(mcp_buffer>=0, msg = "Buffered mcp distance has to be positive!")
+
     # Messenger
     if(getOption('ibis.setupmessages')) myLog('[Setup]','green','Creating distribution object...')
 
     # Convert limits if provided
     if(!is.null(limits)){
+      # Set methods in case a layer was supplied
+      limits_method <- "zones"
       # Convert to polygon if raster
       if(inherits(limits,'SpatRaster')){
         assertthat::assert_that(terra::is.factor(limits),
@@ -147,10 +169,18 @@ methods::setMethod(
 
       # Get fir column and rename
       limits <- limits[,1]; names(limits) <- c('limit','geometry')
-    }
+      limits <- list(layer = limits, "limits_method" = "zones",
+                     "mcp_buffer" = mcp_buffer, "limits_clip" = limits_clip)
+    } else if(limits_method == "mcp"){
+      # Specify the option to calculate a mcp based on the added data.
+      # This is done directly in train.
+      limits <- list("layer" = NULL, "limits_method" = "mcp",
+                     "mcp_buffer" = mcp_buffer, "limits_clip" = limits_clip)
 
-    # Convert to waiver if NULL
-    if(is.null(limits)) limits <- new_waiver()
+    } else {
+      # Convert to waiver if NULL
+      if(is.null(limits)) limits <- new_waiver()
+    }
 
     # Create BiodiversityDistribution object
     bdproto(NULL, BiodiversityDistribution,